requirements.txt
pandas==2.2.2 
mysql-connector-python==8.3.0 
kafka-python==2.0.2 
requests==2.32.3 
python-dotenv==1.0.1 
loguru==0.7.2 
apache-airflow==2.9.1
pymongo
email-validator

prefect==2.14.0
aiohttp==3.8.0

pydantic-settings>=2.0
pydantic>=2.0